+++ src/utils/helpersSensCalc.test.tsx
@@ -23,23 +23,23 @@ import { SUPPLIED_TYPE_INTEGRATION, SUPPLIED_TYPE_SENSITIVITY } from '@/utils/co
 
 describe('Sensitivity Calculator helper functions', () => {
   test('Telescope type, TELESCOPE_LOW', () => {
-    expect(isLow(TELESCOPE_LOW)).toBe(true);
+    expect(isLow(TELESCOPE_LOW)).equal(true);
   });
   test('Telescope type, TELESCOPE_MID', () => {
-    expect(isLow(TELESCOPE_MID)).toBe(false);
+    expect(isLow(TELESCOPE_MID)).equal(false);
   });
 
   test('Numeric check, Returns false when supplied value is "1.4"', () => {
-    expect(isNumeric('1.4')).toBe(false);
+    expect(isNumeric('1.4')).equal(false);
   });
   test('Numeric check, Returns true when supplied value is "14"', () => {
-    expect(isNumeric('14')).toBe(true);
+    expect(isNumeric('14')).equal(true);
   });
 
   test('Beam Size', () => {
     expect(
       getBeamSize({ beam_maj: { value: 1000, unit: 'max' }, beam_min: { value: 0, unit: 'min' } })
-    ).toStrictEqual('1000.0" x 0.0"');
+    ).equal('1000.0" x 0.0"');
   });
 
   test('Transform Surface Brightness Per Sub Band Data, MID continuum default 4 subbands values', () => {
@@ -54,7 +54,7 @@ describe('Sensitivity Calculator helper functions', () => {
       }
     };
     const expected = '1.98 K - 1.50 K';
-    expect(transformSurfaceBrightnessPerSubBandData(minMax)).toStrictEqual(expected);
+    expect(transformSurfaceBrightnessPerSubBandData(minMax)).equal(expected);
   });
   test('Transform Surface Brightness Per Sub Band Data, LOW continuum default 4 subbands values', () => {
     const minMax = {
@@ -68,7 +68,7 @@ describe('Sensitivity Calculator helper functions', () => {
       }
     };
     const expected = '583.47 K - 389.00 K';
-    expect(transformSurfaceBrightnessPerSubBandData(minMax)).toStrictEqual(expected);
+    expect(transformSurfaceBrightnessPerSubBandData(minMax)).deep.equal(expected);
   });
 
   test('Transform Synthesized Beam Size Per Sub Band Data, MID continuum default 3 subbands values', () => {
@@ -95,7 +95,7 @@ describe('Sensitivity Calculator helper functions', () => {
       }
     };
     const expected = `0.47" x 0.42" - 0.33" x 0.29"`;
-    expect(transformSynthesizedBeamSizePerSubBandData(minMax)).toStrictEqual(expected);
+    expect(transformSynthesizedBeamSizePerSubBandData(minMax)).equal(expected);
   });
   test('Transform Synthesized Beam Size Per Sub Band Data, LOW continuum default 3 subbands values', () => {
     const minMax = {
@@ -121,7 +121,7 @@ describe('Sensitivity Calculator helper functions', () => {
       }
     };
     const expected = `7.77" x 6.05" - 2.59" x 2.02"`;
-    expect(transformSynthesizedBeamSizePerSubBandData(minMax)).toStrictEqual(expected);
+    expect(transformSynthesizedBeamSizePerSubBandData(minMax)).equal(expected);
   });
 
   test('Transform Per Sub Band Data, MID continuum default 2 subbands values', () => {
@@ -136,7 +136,7 @@ describe('Sensitivity Calculator helper functions', () => {
       }
     };
     const expected = '88.27 μJy - 73.89 μJy';
-    expect(transformPerSubBandData(minMax)).toStrictEqual(expected);
+    expect(transformPerSubBandData(minMax)).deep.equal(expected);
   });
   test('Transform Per Sub Band Data, LOW continuum default 2 subbands values', () => {
     const minMax = {
@@ -150,7 +150,7 @@ describe('Sensitivity Calculator helper functions', () => {
       }
     };
     const expected = '129.18 μJy/beam - 99.83 μJy/beam';
-    expect(transformPerSubBandData(minMax)).toStrictEqual(expected);
+    expect(transformPerSubBandData(minMax)).equal(expected);
   });
 
   test('Transform Per Sub Band Time', () => {
@@ -164,75 +164,75 @@ describe('Sensitivity Calculator helper functions', () => {
         unit: 'Jy / beam'
       }
     };
-    expect(transformPerSubBandTime(minMax)).toStrictEqual('224.05 μs - 149.68 μs');
+    expect(transformPerSubBandTime(minMax)).deep.equal('224.05 μs - 149.68 μs');
   });
 
   test('Coordinate check, is Galactic, Returns false when supplied value is "RA_TYPE_ICRS"', () => {
-    expect(isGalactic(RA_TYPE_ICRS)).toBe(false);
+    expect(isGalactic(RA_TYPE_ICRS)).equal(false);
   });
   test('Coordinate check, is Galactic, Returns true when supplied value is "RA_TYPE_GALACTIC"', () => {
-    expect(isGalactic(RA_TYPE_GALACTIC)).toBe(true);
+    expect(isGalactic(RA_TYPE_GALACTIC)).equal(true);
   });
 
   test('Supplied Time check, Returns true when supplied value is "SUPPLIED_TYPE_INTEGRATION"', () => {
-    expect(isSuppliedTime(SUPPLIED_TYPE_INTEGRATION)).toBe(true);
+    expect(isSuppliedTime(SUPPLIED_TYPE_INTEGRATION)).equal(true);
   });
   test('Supplied Time check, Returns false when supplied value is "SUPPLIED_TYPE_SENSITIVITY"', () => {
-    expect(isSuppliedTime(SUPPLIED_TYPE_SENSITIVITY)).toBe(false);
+    expect(isSuppliedTime(SUPPLIED_TYPE_SENSITIVITY)).equal(false);
   });
 
   test('Map Sensitivities Units', () => {
-    expect(getSensitivitiesUnitsMapping(3)).toStrictEqual('μJy/beam');
+    expect(getSensitivitiesUnitsMapping(3)).equal('μJy/beam');
   });
 
   test('Shift Sensitivity { value : 50000000000, unit : Jy}', () => {
-    expect(shiftSensitivity({ value: 5000000000, unit: 'Jy' })).toStrictEqual({
+    expect(shiftSensitivity({ value: 5000000000, unit: 'Jy' })).deep.equal({
       value: 5000000000,
       unit: 'Jy'
     });
   });
   test('Shift Sensitivity { value : 40000000000, unit : 3}', () => {
-    expect(shiftSensitivity({ value: 4000000000, unit: '3' })).toStrictEqual({
+    expect(shiftSensitivity({ value: 4000000000, unit: '3' })).deep.equal({
       value: 4,
       unit: 'Jy/beam'
     });
   });
   test('Shift Sensitivity { value : 300000000, unit : 3}', () => {
-    expect(shiftSensitivity({ value: 300000000, unit: '3' })).toStrictEqual({
+    expect(shiftSensitivity({ value: 300000000, unit: '3' })).deep.equal({
       value: 300,
       unit: 'mJy/beam'
     });
   });
   test('Shift Sensitivity { value : 200000, unit : 3}', () => {
-    expect(shiftSensitivity({ value: 200000, unit: '3' })).toStrictEqual({
+    expect(shiftSensitivity({ value: 200000, unit: '3' })).deep.equal({
       value: 200,
       unit: 'μJy/beam'
     });
   });
   test('Shift Sensitivity { value : 100, unit : 3}', () => {
-    expect(shiftSensitivity({ value: 100, unit: '3' })).toStrictEqual({
+    expect(shiftSensitivity({ value: 100, unit: '3' })).deep.equal({
       value: 100,
       unit: 'nJy/beam'
     });
   });
   test('Shift Sensitivity { value : 0, unit : 3}', () => {
-    expect(shiftSensitivity({ value: 0, unit: 'nJy/beam' })).toStrictEqual({
+    expect(shiftSensitivity({ value: 0, unit: 'nJy/beam' })).deep.equal({
       value: 0,
       unit: 'nJy/beam'
     });
   });
 
   test('Time Conversion, In values 100, 3. Out value 6000', () => {
-    expect(timeConversion('100', 3)).toBe(6000);
+    expect(timeConversion('100', 3)).equal(6000);
   });
   test('Time Conversion, In values 100, 3, 2. Out value 1.6666666666666667', () => {
-    expect(timeConversion('100', 3, 2)).toBe(1.6666666666666667);
+    expect(timeConversion('100', 3, 2)).equal(1.6666666666666667);
   });
 
   test('Shift Time, no value', () => {
     const time = undefined;
     const expected = { value: 0, unit: '' };
-    expect(shiftTime(time as any, false)).toStrictEqual(expected);
+    expect(shiftTime(time as any, false)).deep.equal(expected);
   });
   test('Shift Time, value 0', () => {
     const time = {
@@ -240,7 +240,7 @@ describe('Sensitivity Calculator helper functions', () => {
       unit: 's'
     };
     const expected = { value: 0, unit: 's' };
-    expect(shiftTime(time as any, false)).toStrictEqual(expected);
+    expect(shiftTime(time as any, false)).deep.equal(expected);
   });
   test('Shift Time, s to d', () => {
     const time = {
@@ -251,7 +251,7 @@ describe('Sensitivity Calculator helper functions', () => {
       value: 115740.74074074074,
       unit: 'd'
     };
-    expect(shiftTime(time, false)).toStrictEqual(expected);
+    expect(shiftTime(time, false)).deep.equal(expected);
   });
   test('Shift Time, s to h', () => {
     const time = {
@@ -262,7 +262,7 @@ describe('Sensitivity Calculator helper functions', () => {
       value: 833.3333333333334,
       unit: 'h'
     };
-    expect(shiftTime(time, false)).toStrictEqual(expected);
+    expect(shiftTime(time, false)).deep.equal(expected);
   });
   test('Shift Time, s to min', () => {
     const time = {
@@ -273,7 +273,7 @@ describe('Sensitivity Calculator helper functions', () => {
       value: 16650,
       unit: 'min'
     };
-    expect(shiftTime(time, false)).toStrictEqual(expected);
+    expect(shiftTime(time, false)).deep.equal(expected);
   });
   const secondsOnly = true;
   test('Shift Time, s to s', () => {
@@ -285,7 +285,7 @@ describe('Sensitivity Calculator helper functions', () => {
       value: 161.87084064089981,
       unit: 's'
     };
-    expect(shiftTime(time, secondsOnly)).toStrictEqual(expected);
+    expect(shiftTime(time, secondsOnly)).deep.equal(expected);
   });
   test('Shift Time, s to ms', () => {
     const time = {
@@ -296,7 +296,7 @@ describe('Sensitivity Calculator helper functions', () => {
       value: 18.506663149092244,
       unit: 'ms'
     };
-    expect(shiftTime(time, secondsOnly)).toStrictEqual(expected);
+    expect(shiftTime(time, secondsOnly)).deep.equal(expected);
   });
   test('Shift Time, s to μs', () => {
     const time = {
@@ -307,7 +307,7 @@ describe('Sensitivity Calculator helper functions', () => {
       value: 1.8506663149092246,
       unit: 'μs'
     };
-    expect(shiftTime(time, secondsOnly)).toStrictEqual(expected);
+    expect(shiftTime(time, secondsOnly)).deep.equal(expected);
   });
   test('Shift Time, s to ns', () => {
     const time = {
@@ -318,38 +318,38 @@ describe('Sensitivity Calculator helper functions', () => {
       value: 0.1850666314909224,
       unit: 'ns'
     };
-    expect(shiftTime(time, secondsOnly)).toStrictEqual(expected);
+    expect(shiftTime(time, secondsOnly)).deep.equal(expected);
   });
 
   test('Image Weighting, value 0, return natural', () => {
-    expect(getImageWeightingMapping(0)).toBe('natural');
+    expect(getImageWeightingMapping(0)).equal('natural');
   });
 
   test('Image Weighting, value 1, return uniform', () => {
-    expect(getImageWeightingMapping(1)).toBe('uniform');
+    expect(getImageWeightingMapping(1)).equal('uniform');
   });
 
   test('Image Weighting, value 2, return robust', () => {
-    expect(getImageWeightingMapping(2)).toBe('robust');
+    expect(getImageWeightingMapping(2)).equal('robust');
   });
 
   test('Robust mapping, value 1, return -2', () => {
-    expect(getRobustMapping(1)).toBe(-2);
+    expect(getRobustMapping(1)).equal(-2);
   });
 
   test('Robust mapping, value 2, return -1', () => {
-    expect(getRobustMapping(2)).toBe(-1);
+    expect(getRobustMapping(2)).equal(-1);
   });
 
   test('Robust mapping, value 3, return 0', () => {
-    expect(getRobustMapping(3)).toBe(0);
+    expect(getRobustMapping(3)).equal(0);
   });
 
   test('Robust mapping, value 4, return 1', () => {
-    expect(getRobustMapping(4)).toBe(1);
+    expect(getRobustMapping(4)).equal(1);
   });
 
   test('Robust mapping, value 5, return 2', () => {
-    expect(getRobustMapping(5)).toBe(2);
+    expect(getRobustMapping(5)).equal(2);
   });
 });