+++ CHANGELOG.md
@@ -18,6 +18,7 @@ UNRELEASED
 * Fixed: use of REACT_APP_USE_LOCAL_DATA for local development
 * Changed: to show notification when there is an senscalc error after mapping target-observation
 * Changed: Added validation to target fields 
+* Added: Calibration page content
 * Changed: use ska base images `ska-build-node` and `ska-webserver`